First ideas for an easy requirements installer
This is to help out those who don't know their way around python well enough and still want access to the software. 

This will be part of a split process where people can either have a guided process with simple GUIs and step-by-step help, or choose to edit the code, (and future config files), and then run the command via command line.